src/main/webapp/app/entities/parameter-def-type/route/parameter-def-type-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IParameterDefType } from '../parameter-def-type.model';
import { ParameterDefTypeService } from '../service/parameter-def-type.service';

import { ParameterDefTypeRoutingResolveService } from './parameter-def-type-routing-resolve.service';

describe('ParameterDefType routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: ParameterDefTypeRoutingResolveService;
  let service: ParameterDefTypeService;
  let resultParameterDefType: IParameterDefType | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(ParameterDefTypeRoutingResolveService);
    service = TestBed.inject(ParameterDefTypeService);
    resultParameterDefType = undefined;
  });

  describe('resolve', () => {
    it('should return IParameterDefType returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultParameterDefType = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultParameterDefType).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultParameterDefType = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultParameterDefType).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IParameterDefType>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultParameterDefType = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultParameterDefType).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
